fix(howl): host-width SPU transfer source pointer

+++ game/HOWL/HOWL_Bank.c
@@ -122,7 +122,7 @@ int Bank_AssignSpuAddrs()
 		MEMPACK_ReallocMem((sdata->audioAllocSize + 0x7ff & 0xfffff800) + 0x800);
 
 		ret = LOAD_HowlSectorChainStart(&sdata->KartHWL_CdFile,                        // CdLoc of HOWL
-		                                (void *)((int)sdata->ptrSampleBlock2 + 0x800), // destination
+		                                (void *)((uintptr_t)sdata->ptrSampleBlock2 + 0x800), // destination
 		                                sdata->bankSectorOffset + 1,                   // offset of howl
 		                                sdata->numAudioSectors                         // number of sectors
 		);
@@ -187,7 +187,7 @@ int Bank_AssignSpuAddrs()
 			// start transfer
 			SpuSetTransferStartAddr(spuAddrStart);
 
-			SpuWrite((u32 *)((int)sdata->ptrSampleBlock2 + 0x800), (size_t)sdata->audioAllocSize);
+			SpuWrite((u32 *)((uintptr_t)sdata->ptrSampleBlock2 + 0x800), (size_t)sdata->audioAllocSize);
 		}
 
 		sdata->bankLoadStage++;
